% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detect_dependencies.R
\name{detect_dependencies}
\alias{detect_dependencies}
\title{Detect Dependencies}
\usage{
detect_dependencies(path = getwd(), import_functions = import,
  export_functions = export, source_detect = TRUE)
}
\arguments{
\item{path}{A string listing the filepath to search, defaults to the working directory.}

\item{import_functions}{A character vector listing the import functions. This defaults to a
pre-populated list, but you can pass your own list if you only want to
detect certain dependencies.}

\item{export_functions}{A character vector listing the export functions. This defaults to a
pre-populated list, but you can pass your own list if you only want to
detect certain dependencies.}

\item{source_detect}{Logical. Do you want to detect dependencies between R files when one
R script sources another one?}
}
\value{
A dataframe showing the edge list of dependencies between files.
}
\description{
Detect dependencies within files. This is accomplished by
reading the filenames passed to the various input and output R funcitons,
then matching those filenames between the different R files. The assumption
is that all the files imported to an R script are pre-requisites for that
script, and the R script is the pre-requiste for all exported files.
}

