% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_dependencies.R
\name{detect_file}
\alias{detect_file}
\title{Detect file}
\usage{
detect_file(this.file, function_list)
}
\arguments{
\item{this.file}{The file to be searched.}

\item{function_list}{The list of functions to search within.}
}
\value{
A dataframe showing the functions, their filename
    arguments, adn the searched file.
}
\description{
Reads a file, and detects any filename arguments passed
to any of the listed functions.
}

