% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_dependencies.R
\name{graph_dependencies}
\alias{graph_dependencies}
\title{graph_dependencies}
\usage{
graph_dependencies(files, dependencies)
}
\arguments{
\item{files}{A character vector of filenames within the project,
for instance one produced by list.files(recursive = TRUE)}

\item{dependencies}{A dataframe containing the edge list of dependencies
between files. This is the same file which is passed to easy_make(). The
dataframe should have two fields "file" and "pre_req" showing the file
and its immediate pre-requisite respectively. If a file has multiple
pre-requisites, list each dependency as a separate line in the dataframe.}
}
\value{
A DiagrammeR plot showing the dependency graph.
}
\description{
graph_dependencies
}
\examples{
filelist <-  c("analysis/file1.R", "analysis/file2.R", "analysis/markdown.Rmd",
"dep.RData", "Makefile", "maker.Rproj", "man/hello.Rd",
"mtcars.csv", "mtcars.RData", "output.txt", "R/dep_graph.R",
"R/hello.R", "R/make_maker.R")

dependencies <- data_frame(
file    = c("analysis/file2.R", "analysis/markdown.Rmd", "mtcars.csv",
						"mtcars.RData", "analysis/file2.R"),
pre_req = c("mtcars.csv", "mtcars.RData", "analysis/file1.R",
					"analysis/file2.R", "R/hello.R"))
graph_dependencies(filelist, dependencies)
}

