% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_make.R
\name{easy_make}
\alias{easy_make}
\title{easy_make}
\usage{
easy_make(dependencies, render_markdown = TRUE, path = "Makefile")
}
\arguments{
\item{dependencies}{A dataframe containing the edge list of dependencies
between files. The dataframe should have two fields "file" and "pre_req"
showing the file and its immediate pre-requisite respectively. If a file
has multiple #' pre-requisites, list each dependency as a separate line
in the dataframe.}

\item{render_markdown}{If TRUE, the Makefile will render any Markdown files, if FALSE markdown
files will be ignored.}

\item{path}{Destination path of the Makefile.}
}
\value{
A Makefile
}
\description{
Produce a makefile based on
}
\examples{
easy_make(
detect_dependencies(test_dir),
path = "Makefile_test")

unlink("Makefile_test")
}

