% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_dependencies.R
\name{graph_dependencies}
\alias{graph_dependencies}
\title{graph_dependencies}
\usage{
graph_dependencies(dependencies = detect_dependencies(),
  files = list.files(recursive = TRUE))
}
\arguments{
\item{dependencies}{A dataframe containing the edge list of dependencies
between files. This is the same file which is passed to easy_make(). The
dataframe should have two fields "file" and "pre_req" showing the file
and its immediate pre-requisite respectively. If a file has multiple
pre-requisites, list each dependency as a separate line in the dataframe.}

\item{files}{A character vector of filenames within the project,
for instance one produced by list.files(recursive = TRUE)}
}
\value{
A DiagrammeR plot showing the dependency graph.
}
\description{
graph_dependencies
}
\examples{
test_dir <- system.file("test_project", package = "easyMake")
graph_dependencies(
detect_dependencies(test_dir),
files = list.files(test_dir))
}

